backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.6
python-decouple==3.8
Pillow==11.1.0
